changed HP values to int64

+++ TCC.Core/Data/Player.cs
@@ -107,8 +107,8 @@ namespace TCC.Data
             }
         }
 
-        private int maxHP;
-        public int MaxHP
+        private long maxHP;
+        public long MaxHP
         {
             get { return maxHP; }
             set
